fix: move w3c vc validation to v3
