+++ src/3.0/validate/index.ts
@@ -1,0 +1,1 @@
+export * from "./validate";
