+++ src/3.0/wrap.ts
@@ -3,10 +3,11 @@ import { MerkleTree } from "../shared/merkle";
 import { SchemaId, SignatureAlgorithm } from "../shared/@types/document";
 import { WrappedDocument } from "./types";
 import { digestCredential } from "../3.0/digest";
-import { getSchema, validateSchema as validate, validateW3C } from "../shared/validate";
+import { getSchema, validateSchema as validate } from "../shared/validate";
 import { WrapDocumentOptionV3 } from "../shared/@types/wrap";
 import { OpenAttestationDocument } from "../__generated__/schema.3.0";
 import { encodeSalt, salt } from "./salt";
+import { validateW3C } from "./validate";
 
 const getExternalSchema = (schema?: string) => (schema ? { schema } : {});
